Add functional test
